Install the demo delegate on every pipeline unconditionally
The recording switch now lives in Nuke's configuration, so the demo no
longer needs to decide whether to pass a delegate. Renames
DemoDiagnostics to DemoImagePipelineDelegate to match.
